requirements.txt
# Deep Dreams Dependencies - Simple PIL Version

# Core API dependencies
fastapi==0.117.1
uvicorn[standard]==0.37.0
python-multipart==0.0.6
pydantic>=2.0.0

# Image processing - Simple PIL effects
Pillow>=9.0.0

# Templating
Jinja2==3.1.2

# Development tools
python-dotenv>=1.0.0
